exchange int to double value for analysis
git-svn-id: https://bioclipse.svn.sourceforge.net/svnroot/bioclipse/bioclipse2/trunk@10278 fcb5ba71-d80d-0410-8237-ba3920747fcc

+++ _playground_/net.bioclipse.medea/src/net/bioclipse/plugins/medea/core/CMLSpectrumCreator.java
@@ -28,13 +28,13 @@ class CMLSpectrumCreator {
 		/*obtain the mass of the arrayList which contains the ionized fragments*/
 		IMolecule molecule = fragmentTree.getMolecule();
 		IMolecularFormula formula = MolecularFormulaManipulator.getMolecularFormula(molecule);
-		int massI = (int)MolecularFormulaManipulator.getTotalExactMass(formula);
+		double massI = MolecularFormulaManipulator.getTotalExactMass(formula);
 		
 		cmlSpectrum = new CMLSpectrum();
 		cmlSpectrum.setType("massSpectrum");
 		CMLPeakList peakList = new CMLPeakList();
 		
-		FragmentTreeSub fts = fragmentTree.getFragments(massI);
+		FragmentTreeSub fts = fragmentTree.getFragments((int)massI);
 		
 		for(int i = 0; i <= massI ; i++){
 			fts = fragmentTree.getFragments(i);
